+++ actionpack/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Rails 6.0.0.alpha (Unreleased) ##
 
+*   Fix a bug where DebugExceptions errors out when malformed query parameters are provided
+
+    *Yuki Nishijima*
+
 *   Rails 6 requires Ruby 2.4.1 or newer.
 
     *Jeremy Daer*